refactor(auth): extract SCA domain types to sca_models.py (Sprint 4 Track A Block 1)
- Move SCAChallenge, SCAVerifyResult, SCAMethods from sca_service.py to sca_models.py
- Update sca_service_port.py to import types from sca_models (breaks port→adapter cycle)
- Update sca_service.py to import its own domain types from sca_models
- All 112 auth tests remain green; no public contract change
